feat: switch to feature
