+++ packages/base/src/config/FormatSettings.ts
@@ -1,12 +1,7 @@
+import LegacyDateFormats from "../features/LegacyDateFormats.js";
+import type { LegacyDateCalendarCustomizing, LegacyDateFormat } from "../features/LegacyDateFormats.js";
 import { getFormatSettings } from "../InitialConfiguration.js";
-
-type LegacyDateFormat = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9" | "A" | "B" | "C" | undefined;
-type CalendarFormat = {
-	dateFormat: string,
-	islamicMonthStart: string,
-	gregDate: string,
-};
-type LegacyDateCalendarCustomizing = Array<CalendarFormat>;
+import { getFeature } from "../FeaturesRegistry.js";
 
 type FormatSettings = {
 	firstDayOfWeek?: number,
@@ -29,29 +24,15 @@ const getFirstDayOfWeek = (): number | undefined => {
 	return formatSettings.firstDayOfWeek;
 };
 
-const getLegacyDateCalendarCustomizing = (): LegacyDateCalendarCustomizing => {
-	if (formatSettings === undefined) {
-		formatSettings = getFormatSettings();
-	}
+const legacyDateFormats = getFeature<typeof LegacyDateFormats>("LegacyDateFormats");
 
-	return formatSettings.legacyDateCalendarCustomizing || [];
-};
-
-const getLegacyDateFormat = (): LegacyDateFormat | undefined => {
-	if (formatSettings === undefined) {
-		formatSettings = getFormatSettings();
-	}
-
-	return formatSettings.legacyDateFormat;
-};
+const getLegacyDateCalendarCustomizing = legacyDateFormats ? LegacyDateFormats.getLegacyDateCalendarCustomizing : () => { return [] };
+const getLegacyDateFormat = legacyDateFormats ? legacyDateFormats.getLegacyDateFormat : () => {};
 
 export {
 	getFirstDayOfWeek,
 	getLegacyDateCalendarCustomizing,
 	getLegacyDateFormat,
 };
-export type {
-	FormatSettings,
-	LegacyDateCalendarCustomizing,
-	LegacyDateFormat,
-};
+
+export type { FormatSettings };
